[REF] test_resource: unpack test_resource
- break the test calsses each to individual file
