+++ addons/test_resource/tests/__init__.py
@@ -1,5 +1,9 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
-from . import test_resource
+from . import test_calendar
+from . import test_mixin
 from . import test_performance
+from . import test_resource_errors
+from . import test_resource
+from . import test_timezones
